+++ src/NovelSpider.Config/NovelSpider/Config/Configs.cs
@@ -72,7 +72,7 @@ public class Configs
 		HaveFunction = "ZhanQunPinyinDir中译英";
 		UserID = Guid.NewGuid();
 		AssemblyVersion = Assembly.GetExecutingAssembly().GetName().Version;
-		DisplayVersion = "10.18.5-net10-test";
+		DisplayVersion = "10.18.6-net10-test";
 		TaskNovelInfo = new Dictionary<string, object>();
 		IsDemo = false;
 		Build = new DateTime(2026, 8, 14);